Refactor tracker client
The idea here was to make the tracker client more tolerant of partial responses. I'm not sure the complexity is worth it for such a small class though. What I really want is a way to call recv with fixed windows (say, 1024) but not block when the socket doesn't have that much data on it. I'm not sure how to do this yet.